docs(icon): update icon docs and structure
