+++ .storybook/main.js
@@ -47,7 +47,8 @@ module.exports = {
     '../projects/canopy/src/lib/hero/hero.stories.ts',
     '../projects/canopy/src/lib/hero-img/hero-img.stories.ts',
     '../projects/canopy/src/lib/hide-at/hide-at.stories.ts',
-    '../projects/canopy/src/lib/icon/icons.stories.ts',
+    '../projects/canopy/src/lib/icon/docs/guide.stories.mdx',
+    '../projects/canopy/src/lib/icon/docs/icons.stories.ts',
     '../projects/canopy/src/lib/link-menu/link-menu.stories.ts',
     '../projects/canopy/src/lib/modal/modal.stories.ts',
     '../projects/canopy/src/lib/page/page.stories.ts',
