+++ projects/canopy/src/lib/icon/docs/guide.stories.mdx
@@ -1,0 +1,102 @@
+import { Meta } from '@storybook/addon-docs';
+
+<Meta title="Foundation/UI icon/Guide" parameters={{ viewMode: 'docs', previewTabs: { canvas: { hidden: true } } }} />
+
+# UI icon
+
+<p class="standfirst">UI icons are visual depictions of an object, action or idea. They're simple and easy to recognise, and can communicate a lot with very little.</p>
+
+![example](docs/ui-icon/example.png)
+
+## Usage
+
+UI icons should be used to highlight or support key information or give affordance to links and buttons.
+
+Import the icon module into your module:
+
+```js
+@NgModule({
+  ...
+  imports: [ ..., LgIconModule ],
+})
+```
+
+
+
+Import the `LgIconRegistry` service and register your icons inside your module:
+
+```js
+// import the desired icon
+import { lgIconEmail } from '@legal-and-general/canopy';
+
+export class AppModule {
+  // register the icon using the `LgIconRegistry` service
+  constructor(private iconRegistry: LgIconRegistry) {
+    this.iconRegistry.registerIcons([
+      lgIconEmail
+    ]);
+  }
+}
+```
+
+
+
+and in your HTML:
+
+```html
+<lg-icon
+  name="email">
+</lg-icon>
+```
+
+## Dos and Don'ts
+
+### Do
+
+![do](docs/ui-icon/do.png)
+
+1. **Do** scale icons along the 4px scale to ensure crisp outlines
+2. **Do** use charcoal or high contrast colour icons on white background
+3. **Do** use tinted colours on light colour backgrounds, ensuring a minimum of 4.51:1 contrast
+4. **Do** use white icons on a high contrast background
+
+### Don't
+
+![dont](docs/ui-icon/dont.png)
+
+1. **Don't** scale icons below 16px
+2. **Don't** scale icons above 64px. Use brand icons or illustration for more prominent graphics
+3. **Don't** overlay icons on photographic backgrounds
+4. **Don't** colour the icon differently to the background e.g. yellow on blue
+5. **Don't** use low-contrast colour combinations
+
+## Additional developer details
+
+### Inputs
+
+| Name   | Description          |  Type  |  Default  | Required |
+| ------ | -------------------- | :----: | :-------: | :------: |
+| ``name`` | the name of the icon | string | undefined |   yes    |
+
+All icons have height and width equal to 1em. This means the icons will be as big as the font-size specified by the parent.
+
+By default the icons have the `fill` css property set to `currentColor`.
+
+> `currentColor` acts like a variable for the current value of the `color` property on the element. And the Cascading part of CSS is still effective, so if there’s no defined `color` property on an element, the cascade will determine the value of `currentColor`.
+>
+> *- <a href="https://www.digitalocean.com/community/tutorials/css-currentcolor" target="_blank">Understanding the currentColor Keyword in CSS</a>*
+
+Please note that the `need-help` and `question-mark` icons currently don't have the ability to change colour.
+
+---
+
+## Related
+
+- [Brand icons](./?path=/docs/components-brand-icon--standard-brand-icons)
+
+
+## Help improve this page
+
+To help make sure this page is useful, relevant and up to date, you can <a href="https://forms.office.com/r/6UMyXFeupH" target="_blank">suggest an improvement</a>.
+
+Please take a moment to <a href="https://forms.office.com/r/ryiKP7ygUK" target="_blank">rate your experience of using Canopy</a>.
